adjust int types

+++ entity/user.go
@@ -3,7 +3,7 @@ package entity
 import "github.com/google/uuid"
 
 type User struct {
-	UserID           int
+	UserID           int64
 	UserUUID         uuid.UUID
 	UserExternalUUID string
 	UserName         string
